requirements.txt
huaweicloudsdkall
python-dotenv
mysql-connector-python
requests 